Fix how counters are emitted to Datadog
When implementing the emitter, I chose not to send incremental counter
updates, which changed the behavior of counters and made them less
useful. This restores the default Statsd behavior of using incremental
counters.

Also add tests for the new behavior and explain it in the package docs
to avoid confusion in the future.