Use functional gauges for runtime metrics (#5)
There's no reason to report these at a higher rate than consumers read
them from the registry and switching to functional gauges lets us remove
a goroutine. While making this change, refactor the server code a bit to
avoid having a metrics-specific initialization function, which always
felt weird.

Note that this is a breaking change for users that use the WithMetrics
parameter directly, as it no longer takes an interval. I think this is
acceptable as it's an easy fix and all our existing internal clients use
the default parameters.