Add option to disable encrypted SAML assertions
While you can override this in most IdPs, by default, the inclusion of
an encryption certificate in metadata suggests that the application
requires encrypted assertions. This is a reasonable default, so we
retain that, but add an option to disable this for debugging and in
situations where the claim data is not sensitive.