Add appmetrics package
This package contains utilities that make it easier to construct and
register metrics structs because you only have to specify the metric
name and type in one place. It should hopefully help standardize how we
configure metrics in various applications.

One disadvantage is that it doesn't handle tagged metrics right now,
since the metric names change based on the tag values. I'll have to
think about how to implement this.