Simplify tagged metrics
I realized that the Tagged[M] interface could work after all because we
can get the type parameter from the return type of the Tag() method. We
have to implement our own interface check since different instantiations
of the interface are different types, but this isn't too bad for a
single method.