Add the appmetrics package (#182)
This package contains utilities that make it easier to construct and
register metrics structs, as you only have to specify the metric name
and type once in the struct definition. It should help standardize how
we configure metrics in our applications.

It supports all the standard metric types from rcrowley/go-metrics and
supports tagging metrics using the format supported by the datadog
emitter in this module.

See the documentation and examples in this commit for details on all of
the features and how to use this package to define metric structs.