Add prometheus metrics emitter
The core of this package is a prometheus.Collector implementation that
exposes the metrics of an rcrowley/go-metrics registry. It also provides
a convenience function for creating a collector, registering it with a
Prometheus registry, and exposing it on a HTTP endpoint. If you need to
customize any of the Prometheus options, you can use the Collector
directly.

I ended up writing my own emitter because the existing exporters I found
had some limitations:

* Don't understand our custom tag/label format
* Report at fixed intervals intead of implementing Collector
* Export histograms incorrectly